Add sam_format_aux1 and bam_aux_get_str APIs. (PR #1134)
sam_format_aux1 is extracted out from sam_format1_append, and
indeed is the lion's share of this function.  It is useful as an external
API for anything that is generating its own tags or a partial tag list in
SAM format, such as "samtools fastq".  The tag key, type and data
are passed in as separate pointers to aid usage with bam_aux_get
which returns 2 bytes into the in-memory representation of a tag.
It also allows the function to be used in cases where these are not
stored together, for example the internal CRAM representation.

bam_aux_get_str is a wrapper around sam_format_aux1 which
simplifies the case of converting a single tag from a record in
a bam1_t struct.