The pileup constructor now handles errors.
The API for this always had both constructor and destructor as type
int, but it was never documented what this return value is for and nor
was it ever checked.

Although the ABI hasn't changed, this is an API change as we have
added meaning to an unused element.  However (hopefully) the new
meaning is in line with expectations and library norms.