Use the sover-bumping version number for -compatibility_version
It seems that -compatibility_version checking has to date been ineffective
on macOS, which could explain why it seemed that compatibility_version was
not intended to be compared to current_version. Switch to a more orthodox
scheme of using the version number of the HTSlib release that bumped the
soversion as the compatibility_version.

The Makefile now hard-codes LIBHTS_SOVERSION and LIBHTS_SOBUMP_VERSION,
both of which should be updated when the soversion is bumped.