Replace the indexing range check error message.
This code detects when the end positioni s too large for the current
index.  It then computes an adjusted n_lvls parameter and tells the
user to switch to CSI with this parameter.

That has numerous failings though.

1. The only user-adjustable parameter is min-shift, not levels, so
the advise is impossible to follow.

2. CSI auto-scales based on SQ headers, so we don't need to be
explicit anyway (assuming we currently have a BAI index that doesn't
fit).

3. If we already were using CSI index and if we did have the
configuration options to do what it asks, it wouldn't fix the problem.
Given CSI auto-scales, if we see this message it's because of the
alignments don't match the headers; either the header has been
replaced by something inappropriate or there are bugs in the aligner
that has emittede rogue POS fields.  There is no fix to the index that
actually fixes the data; it's simply brushing the problem under the carpet.