Replaced non-standard compliant check for a pthread being initialized with a more standards compliant mechanism.
The problem was this test (sam.c line 2343): if (/.../ && fd->dispatcher)
According to the pthreads standard, the pthread_t type should be considered opaque.
On some platforms it is implemented as a struct rather than an integer, causing a compilation error.
In addition, if a pthread_create call fails, the value of the thread identifier is undefined.