Further work on fastq writing, plus some read updates.
It now uses the fastq_state struct to track options and tidied up
the code a bit.

Also added AUX tag output to fastq_format1 and made both input and
output tag lists filterable.  By default it's all tags, but it can be
specified as a specific comma-separated tag whitelist.  This matches
the behavior of "samtools fastq -T".

Added tests for writing.

Replace use of kseq by own dedicated fastq parser (reader)

Permit the fastq barcode tag to be adjusted (both reading and writing).
This is for both parsing CASAVA tags into aux tags, and back out again.

Permit fastq mode 'f' to be used in conjunction with bgzf mode 'z'.

FASTQ_OPT_AUX string is now processed with a hand written tokeniser
instead of strtok_r.  Strtok is just more trouble than it's worth!
The input needs to be writeable, meaning we cannot pass in literal
strings.

This string is now also tokenised to 16-bit integers (2 chars of tag
code) rather than as a 2 byte string.  This means the khash used
doesn't require the argument passed in to be resident in memory for
the duration (or to have to strdup it itself to compensate for this).

Finally the CASAVA tag writing now has a check for non "+" symbols
between barcodes.  The SAM spec recommends "-", so we spot and alter
these.