Permit FASTA read and write as sequence_data.
This is prompted by wanted to simplify "samtools fastq", or more
specifically the "fasta" variant of it (it's the same code).

I don't want to have to keep all the old tag formatting cruft in
samtools/bam_fastq.c, especially as it has a few buglets, so that
means we want htslib to be able to write in fasta format if we're to
keep both "samtools fastq" and "samtools fasta".

Logically speaking we should be symmetric, permitting reading fasta as
well as writing fasta.  However this may not necessarily be what
people were expecting... Time will tell whether it gives rope for
people to hang themselves with, eg "samtools view GRCh37.fa".  (That
works, but it's more likely they were wanting to specify a reference
to use along side some command reading from stdin and forgot the -T.)

Plus more checking of file suffixes and gzip options.

I can't believe I've found 5 separate functions that recognise file
formats and act on them.  A bit of a tangled web, but I believe this
is now all of them recognising fastq/fasta properly.