Fix appveyor.
The keys used to sign MSYS package have changed, so we have to
update the key store first.

The next problem is that newer MSYS packages are compressed with zstd,
but the installed pacman cannot handle these.  We need to explicitly
update pacman first.

Then finally we can install our build system.

Note none of this is strictly necessary.  The alternative solution is
to comment out all pacman updates as the base image has sufficient
compilers and libraries to build and pass the test harness.

The main difference is the lack of libcurl, which then means we don't
test compilation of the components that depend on it.  We don't test
these anyway, but it's useful to know they build at least.  However if
we hit problems in the future due to the explicit path of those key
files then that's a viable temporary workaround.

(It may be that AppVeyor also updates their base image at some point,
in which case the key shenanigans can be deleted.)