Fix a long-standing typo in hts_lrand48.
The return type of the definition was wrong, causing callers
using the function to get incorrect results in some cases,
mainly depending on the platform and compiler options used.

Note: this function is not used by htslib.  It was only added for
reasons of completeness when we were putting in hts_drand48, used
by htslib/ksort.h.