Don't link plugins with libhts.dylib on macOS
PR #1072 changed plugin linking so that plugins are linked back to
the dynamic libhts.so/.dylib, to facilitate use when libhts is itself
dynamically dlopen()ed with RTLD_LOCAL, e.g., by the Python runtime
which uses default dlopen() flags which on Linux means RTLD_LOCAL.

This broke plugin loading on macOS when opening plugins in an executable
in which libhts.a has been statically linked, as there were then two
copies of the library globals (notably hfile.c::schemes), one from the
executable's libhts.a and one from the plugin's libhts.NN.dylib.
(The Linux loading model does not suffer from this issue.)

The default dlopen() flag on macOS is RTLD_GLOBAL, so this can be fixed
by reverting the change (on macOS only) and depending on the symbols
supplied by a static libhts.a, a dynamically linked libhts.NN.dylib,
or a RTLD_GLOBALly dlopen()ed libhts.NN.dylib. This rebreaks the case
of dlopen()ing libhts on macOS while explicitly specifying RTLD_LOCAL,
but this is not a common case. Fixes #1176.

Disable the `plugins-dlhts -l` test case on macOS. Add a test of
accessing plugins from an executable with a statically linked libhts.a
(namely, htsfile) to test/test.pl.