Add a general SAM/BAM/CRAM input filter option (HTS_OPT_SAM_FILTER).
This uses an expression language and SAM format fields to specify
rules which must pass (so filter-in rather than filter-out).

The code structure is expr.[ch] implements a generic expression
language which accepts a function callback for performing variable
lookup, with sam.c:bam_sym_lookup() being that function.

The sam_read1 function has been changed to keep looping until the
filter passes, or we run out of data.

This works for all file formats, but does so by adding a new member to
htsFile.  Technically it's an ABI change, but at least our own usage
of it never uses arrays of them and it's stated in the header to be an
opaque type which in theory should also ban that, albeit without any
compiler checks).  If this is a problem, would could add the same
member to BGZF, cram_fd and hFILE as they're all private, but that
feels like needless hoop jumping.

Syntax supports

    Grouping:      (, ),   eg "(1+2)*3"
    Values:        integers, floats, strings or variables
    Unary ops:     +, -, !, ~  eg -10 +10, !10 (0), ~5 (bitwise not)
    Math ops:      *, /, %  [TODO: add // for floor division?]
    Math ops:      +, -
    Bit-wise:      &, |, ^  [NB as 3 precedence levels, in that order]
    Conditionals:  >, >=, <, <=,
    Equality:      ==, !=, =~ !~
    Boolean:       &&, ||

Data types are numerics (integers or doubles, eg 10, -0x2B, 10.2e-7)
and strings "str" with backslash as an escape char within strings.

The key-words currently supported are:

    flag                    int
    flag.paired             int
    flag.proper_pair        int
    flag.unmap              int
    flag.munmap             int
    flag.reverse            int
    flag.mreverse           int
    flag.read1              int
    flag.read2              int
    flag.secondary          int
    flag.qcfail             int
    flag.dup                int
    flag.supplementary      int
    mapq                    int
    mpos / pnext            int
    mrefid                  int
    mrname / rnext          string
    ncigar                  int
    pos                     int
    qlen                    int
    qname                   string
    refid                   int
    rlen                    int
    rname                   string
    tlen                    int
    [XX] tag                int / string

For example, to return only records where the mate reference name
differs to the mapped query name and the read is aligned with high
mapping quality:

    test_view -i 'sam_filter=mqual >= 30 && mrname != rname' in.bam

Filters to find significant deletions:

     cigar =~ "[0-9]{2}D"
     rlen - qlen > 10

Duplicates that aren't part of a "proper pair" (equivalent to
"samtools view -F 2 -f 0x400")

    sam_filter=flag.dup && !flag.proper_pair