Filter "qlen" tweak, add seq/qual/library, and add functions.
"seq", "qual" and "library" are string variables, so we can now match
them with e.g. seq =~ "(AT){10}".

We have basic functions that operate on strings, interpreting them as
arrays of characters.  (This could be updated for B aux arrays and
larger objects in the future.)  Functions are length, min, max and
avg.  For example avg(qual) or length(seq).  Note qual here is binary,
so NOT the ASCII phred+33.  I wouldn't advise trying to match it in a
regexp.

"qlen" was previously length of sequence, and not the value of
cigar2qlen.  This differs when we have CIGAR but seq is "*".  Use the
cigar2qlen definition so it is symmetric with our "rlen" variable.

To permit the previous definition, we can now do "length(seq)".

Note that min and max of empty string is NAN.  This isn't treated as
an error.  This permits us to do "avg(qual) > 10" where some entries
have seq and qual both as "*" (ie zero length).