Further optimisations to probaln_glocal.
The changes are:
- Replace a lot of the repeated set_u macro usages that go from k to
  {u, v01, v10, v11} with pointer increments instead.  This rather
  obfuscates things as the difference between v01/u and v11/v10 (+/-3)
  is now hidden (but see PROBALN_ORIG comment below).

- An attempt at manual ordering of code to avoid as many instruction
  latency issues as possible.  Mostly compilers do this already, but
  some combinations benefit here.

- Cache and reuse of variables computed during the previous loop
  iterations.  This can have a big impact, but again whether or not it
  was already done for us is compiler specific.

- Removal of some loop invariants (ideally already done for us, but
  safer to do it ourselves).

- Perform the bi (backwards) rescaling during the backwards
  calculation loop instead of in a second pass.

- Reduce conditionals in the map stage.

- Plus some code formatting tidyups, in particular removal of
  many;statements;on;same;long;line.

The original code wasn't exactly easy to understand, but the new code
is possibly even worse if that's possible.  Hence I've kept the old
code in place guarded by a #ifdef PROBALN_ORIG (technically not
original as bdf85e4 already did some optimisations) to act as a bit of
a translation guide.  If we prefer to simply keep this in git history
then feel free to nuke the ifdefed bits.

Benchmarks vs PROBALN_ORIG (time in this function only)
    35% quicker with gcc 7 -O2     (NEW: 1168742 usec, ORIG: 1575843)
    14% quicker with gcc 7 -O3     (NEW: 1165683       ORIG: 1333712)
    33% quicker with gcc 9 -O2     (NEW: 1169097       ORIG: 1556956)
    29% quicker with gcc 9 -O3     (NEW: 1157547       ORIG: 1491839)
    24% quicker with clang 7.0 -O2 (NEW: 1207968       ORIG: 1501549)
    24% quicker with clang 7.0 -O3 (NEW: 1211660       ORIG: 1504905)

On samtools mpileup of 10 million reads this reduced the time from
6m58 to 5m27 (28% faster throughput).  This is because 86% of all CPU
time was spent in this one function!  It's still 81.5% even after
optimisation.

Note I initially attempted a SIMD implementation, which I think would
still be possible, but it cannot be faster without majorly changing
the data layout and order of evaluation, plus realistically it'd need
AVX or above as SSE4 can only do SIMD on 2 doubles so the overhead is
significant.