Split cram_block_method enum into public and private.
Also the public version has had the enum members renamed to include an
HTS_ prefix.  At present they appear to be unused, although it's hard
to know for sure with some of them due to the common hits you get all
over github.  Leaving them as they are though is just waiting for a
hard problem to arrive.

This commit introduces the notion of HTS_COMPAT.  The ABI changes
changed, but the API has (albeit in a way we believe should affect
people).  It's a 3+ digit number with the last two being minor version
and earlier being major.  So htslib version 1.12 would be 112.

The intention is simply to permit packages with dependencies on
specific earlier API names to still compile with the minimal of
changes (even if it's just by overriding CFLAGS).