SQUASH ME: Add back public API for knetfile.h.
Plus include stub knet functions for ABI compatibility.

The knetfile.h has HTS_DEPRECATED markers for the knet_* functions, so
anyone compiling against the new code will get a warning that we
intend to delete these in the future.  It's debateable though whether
it should exist at all.  We could just remove this header right now
and break API, but retain the previous ABI.

The stub copy in knetFile is now divorced knetfile.[ch] and has its
own implementation which simply wraps up hFILE instead.

The knet_tell macro does direct access of knetFile::offset so we
cannot simply wrap up htell as we'd like.  Similar knet_fileno
directly queries knetFile::fd.  However the only use I've found of
this was in old copies of bgzf and bcftools main which then did fstat
on it so it's only filled out for local files.  (That's good because
getting it out of curl isn't supported by our code.)