Add configure check for the htscodecs submodule
So that anyone who tries to build the library without setting
up the submodule first will get a hint on how to make the build
work.

The check isn't really needed for anyone building from a release
tarball, but it will be harmless as it should always succeed (as
long as the tarball was built correctly and includes the htscodecs
files).