Add Makefile check for the htscodecs submodule
Currently if the submodule files are missing it asks the user
to run "git submodule update" to install them.  It's possible
that the Makefile run this itself and continue, although
I'm not sure if that would be completely safe in parallel builds.

The variables lising dependencies for htscodecs header files
are moved to htscodecs_bundled.mk to ensure they are only set
for submodule builds.