Fix incorrect setting of end position in overlap_push()
Commit a2fdf3b changed bam_endpos() so that it would consider
alignments which consume no bases (e.g. fully soft clipped) as
one base long.  This had the side effect of making overlap_push()
change the position stored in a->end, causing reads that would
previously have been ignored to be added to the pileup.  As
resolve_cigar2() was not designed to handle such reads it
resulted in nonsensical values being set in the pileup structures
for these reads.

The update to a->end in overlap_push() is not necessary anyway
as it will have already have been set correctly in bam_plp_push(),
so the simplest solution is to remove the call to bam_endpos()
in overlap_push().

Pointers to the cigar data in tweak_overlap_quality() have been
changed to const to ensure it makes no changes that would change
the alignment end position.

Fixes samtools/bcftools#1362 (bcftools mpileup seg fault for reads
with only 1 match).