Renumber introspection feature codes (and they are unsigned)
Leave gaps in the allocated bits between the different feature
categories. Leave a gap at the end for HTS_FEATURE_LIBS if it is
added in future. Don't mention htslib_plugin_path() in the public
header as it is an internal function.

Change htslib_test_feature() to use unsigned int, as per htslib_features().