remove a strange gcc10 warning.
Gcc 10 when using -O and -fsanitize=address produces a warning about
mallocing some insanely large amount of memory.

This is an impossibility given the limits on the sizes of some data
values, but gcc doesn't know this and is assuimng variables may wrap
around and go negative.  Using an unsigned type stops it from
exploring the inaccessible data ranges.