Fix undefined behaviour warning in kstring
Fixes a "runtime error: applying zero offset to null pointer"
warning from clang 10.0.1 undefined behaviour sanitizer.

This happened when the storage for the string had not been
allocated before calling kvsprintf().  Fix by making it allocate
a buffer if this is the case, so vsnprintf() has something to
write into.