Fix undefined behaviour in bcf_update_alleles()
bcf_update_alleles() checked to see if any of the allele pointers
passed in were pointing to the existing allele data, so it could
avoid clobbering them during the update.  This caused a "runtime
error: applying zero offset to null pointer" from undefined
behaviour sanitizer when it was called with a new BCF structure
where line->d.als had not yet been allocated.  However, even
when it had been allocated, the test relied on undefined
behaviour as it compared pointers to different memory regions.

Fix by rewriting the function so that it always copies the
input data to a new location.  Where the new alleles are short
enough (which should be the usual case) and can fit in the
existing allocation, they are first copied into a stack buffer,
and then back to line->d.als; otherwise they go into a newly
allocated buffer which is switched for the original.  This avoids
the undefined behaviour issues, and also allows the line->d.als
memory to be reused even when some of the input alleles point
into it.

Includes unit tests.