Change MACH_O compatibility and current versions
When building libhts.dylib, the  compatibility_version and
current_version were not being set in the way documented by
Apple - notably they expect current_version to be from the
same sequence as, and greater than compatibility_version. See:
https://developer.apple.com/library/archive/documentation/DeveloperTools/Conceptual/DynamicLibraries/100-Articles/DynamicLibraryDesignGuidelines.html

Change the values so current_version is related to the current
HTSlib version, and compatibility_version the last one that
introduced a backwards compatible change.  As earlier HTSlib
releases set compatibility_version to 3, we prepend '3.' to
these numbers so programs linked to earlier versions will still
work with libraries built after this change.  This will be removed
on the next ABI-changing update to the library, which will allow
the compatibility_version and current_version sequences to be
reset.

Thanks to John Marshall for help with this issue.