Update to htscodecs release 1.0
Updates the htscodecs submodule and updates dependencies for the
new header files.

Adds Makefile infrastructure to build and clean up the
htscodecs/htscodecs/version.h file, for git checkouts only.
Release tarballs will ship with a pre-built copy of this file.

Fixes issue where the Makefile rules to test if the submodule is
missing would cause missing source files to be silently ignored.
They now run a recipe to check for the target, and suggest
updating the submodule if it's missing.

Adds documentation for the '--with-external-htscodecs' configure
option to INSTALL.

The hts_version() function is unchanged reporting pure htslib only.
Use hts_test_feature(HTS_FEATURE_HTSCODECS) to query htscodecs version.