Recognise legacy RAZF compression
RAZF is an obsolete predecessor to BGZF, and is similarly a variant of
GZIP using an extra header field. It also adds a trailing index table.

Adding this htsCompression value does not affect bgzf_read_init()'s
detection of BGZF vs plain-GZIP; RAZF remains treated as is_gzip and
the trailing index table is not handled well, leading to problems
if you try to decompress such a legacy file with e.g. bgzip -d.