Refuse to read the legacy RAZF compression format
Instead emit an error message recommending the use of gunzip to decompress
the file, in the unlikely event a RAZF file is encountered. If seeking is
available, attempt to read the sizes stored at the end of the RAZF trailing
index table so that the message can show a truncate command to remove the
index table before gunzipping the file.