Permit platform specific BAQ parameters.
The flag passed to sam_prob_realn can now contain platform data,
encoded along with other bits via the htsRealnPlatform enum.

It also automatically selects long-read parameters if the read length
is bigger than 1kb, but otherwise it defaults to the same parameters
as before.

This helps bcftools mpileup call SNPs on PacBio CCS reads.

TODO: ideally we would have an additional API that accepts a BAM
header and returns a platform field for us to ease use of this
function. Eg:

   sam_prob_realn(b, ref, len, BAQ_APPLY | BAQ_EXTEND | sam_baq_platform(hdr));

Or to add a new function taking an extra argument for the tuned
probaln_par and similar header-processing function. Eg:

   sam_prob_realn2(b, ref, len, BAQ_APPLY | BAQ_EXTEND,
                   sam_platform_probaln_par(hdr));