Remove compressBound assertions. (PR #1258)
These trip on with zlib-ng as the worse case expansion is 9-bits per
byte (at level 1 only).  The test is done on opening a file, which
seems strange.

We could do the check on opening for write only, but it seems more
productive to error only when we actually can't fit the data rather
than just incase it may if given truely random input.  Our real data
doesn't seem to trigger this.

Random data with zlib-ng can cause data expansion beyond 64k.  However
bgzf_compress already has code to handle uncompressed blocks, so we
just fall back to that (thanks to John Marshall for the idea).

Fixes #1257