Support building in a separate build directory
Traditional autoconf-using projects generate a directory tree
of recursive Makefiles from a directory tree of Makefile.in files.
Thus when configuring for an out-of-tree build, typically the tree
of builddirs and their Makefiles are created fairly automatically.

HTSlib instead uses a single non-recursive non-generated Makefile that
includes fragments generated by configure. Hence configure.ac needs
to detect out-of-tree configuration as a special case and explicitly
create Makefile and the tree of subdirectories in the build directory.

Add builddir_vars.mk.in, which for an out-of-tree build (only) is
generated into $builddir/htslib_vars.mk. It sets VPATH and some other
additional variables and then includes the real $srcdir/htslib_vars.mk.
This is used to override $srcdir etc for out-of-tree builds, and notably
adds an -I option for htscodecs/version.h which in an in-tree build is
found by being in the same directory as the C file that includes it.

HTSlib also provides htslib.mk for use by external projects. Split its
variables and rules into $(HTSDIR) and $(HTSSRCDIR) as appropriate.
These are the same by default, but builddir_vars.mk.in overrides
$(HTSSRCDIR) to be $(HTSDIR)/$([HTSlib's ]srcdir) instead, taking
care to handle the case of $(srcdir) being an absolute path.

The `include` in htscodecs.mk.in would need to be $(srcdir)/selected.mk
for Makefile but $(HTSPREFIX)selected.mk for htslib.mk. Rather than
arrange this, it's easier to generate $builddir/htscodecs.mk as a symlink
to $srcdir/selected.mk instead, using AC_CONFIG_LINKS.

Add $(srcdir)/ or $(srcprefix) (for tidiness in in-tree build logs) to
commands in recipes containing explicit filenames in source directories.
This has been done for `make all install` but not yet for `make test`.