Prevent assert() in bcf_sr_set_regions
It seems bcf_sr_set_regions ought be used before bcf_sr_add_reader.
However, if a user unaware of this calls bcf_sr_set_regions **after**
bcf_sr_add_reader, they are met with a failing assert. This is odd, as
bcf_sr_set_regions already contains a useful error message about exactly
this error.

This patch reorders the check for incorrect usage and the assert(), so
that if an API user stuffs up the order of the calls, they get the nice
message that already exists and not an assert().

I'm not sure if the assert() is even needed here, and probably is best
replaced with a nice hts_log_error() in any case. I think
`readers->regions != NULL` only when either bcf_sr_add_reader or
bcf_sr_set_regions has already been called. I'll leave this in the
maintainer's hands.