Prevent endless aux_parse loop and avoid malformed aux data
In lenient mode, skip the check for integer overflows.  Fixes an
endless loop due to `overflow` not being reset, and by the time
the check is made all the input data has been used so trying to
loop back won't do anything useful.

Make the _parse_err() macro in aux_parse() reset `b->l_data` to
where it was at the start of the loop if it detects an error in
lenient mode.  This cleans up any semi-formed aux records that
may be present before the parser tries to skip to the next tag.