Fix to make samtools cat work on CRAMs again.
Samtools cat utilises low level APIs to turn codecs from a read
cram_fd to codecs sutiable for a write cram_fd.  This had a missing
initialisation in the HUFFMAN code and a small memory leak, caused by
the addition of the CRAM v4 code (but affecting CRAM v3).

Also while adding CRAM v4, the EOF block was changed from being a
literal string of bytes into simply encoding an empty container and
letting the code auto-generate the correct byte stream.  There were
bugs in cram_encode_compression_header for empty blocks, which
affected cram blocks that had gone through this read/write transcode
process.  Specifically the compression header's num_record check is
invalid as this is only populated for CRAM v1.0.  We check the
container instead, which is always correct and permits us to find the
empty EOF blocks.

Note this EOF writing change also affected more than samtools cat.
All CRAM files being output contained a slightly larger EOF block,
containing an unused preservation map.  It is debateable whether this
is within the spec or not.  The spec states both the meaning of the
EOF block (an empty container with a specific ref seq id and alignment
start) as well as listing what this encodes to in hex.  The former
worked, but the latter differed.

Fixes samtools/samtools#1420