Fix undefined behaviour in cram_xdelta_encode_char.
"dat" is a malloc result and cp_end is computed from dat, irrespective
of whether dat returned NULL or not.  As pointer arithmetic with a
NULL pointer is undefined, this is bad.  We do check for dat being
NULL, but the undefined behaviour has already been triggered.

It is not believed however that this leads to any optimisation bugs or
bounds checks being removed causing security issues.  The compiler
knows malloc *may* (and probably will) return non-NULL so it is
required to produce code for handling that case.  If it chooses to
produce another code path for the NULL case it can, and that could
have optimisation differences, but it'll be unreachable due to the
malloc return value check.