Guard against a gcc 11.1.0 false positive warning.
Gcc 11 complains (it's a warning, but upgraded due to -Werror):

        cram/cram_codecs.c:1916:19: error: 'cp_end' may be used uninitialized [-Werror=maybe-uninitialized]
         1916 |             cp += c->vv->varint_put32(cp, cp_end, zigzag16(c->u.e_xdelta.last));
              |                   ^~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~

The cause of this appears to be because cp (dat) is malloced rather
than calloced, so gcc knows the buffer is uninitialised.  varint_put32
prototype takes "const char *" for the cp_end argument, which leads
gcc to warn because it thinks there is no way the function can modify
the data and it can detect we haven't initialised it before hand.

However it hasn't acknowledged that cp and cp_end are both pointers to
the same malloc block, and the cp is accepted as "char *" by the
varint_put32, so it is indeed writing and not reading.

This has been filed as
https://gcc.gnu.org/bugzilla/show_bug.cgi?id=100417

This commit explicitly removes this warning for this one function on
gcc 11.1 only.  It's included as a demonstration of fixing the issue
incase other people report it, but I am not convinced we should be
doing so as hopefully it will be transient.  (Or, quite possibly, the
gcc team educate us into reasons why our code is undefined in some
manner.)