Silence a gcc 11.1 warning with kseq.
The new uninitialised memory checking in gcc 11 is tripped up as it
thinks there may be overflow cases in ks_expand meaning the resized
buffer still isn't large enough.  We have already fixed this in
kroundup.h, but it's not sufficient to silence it.

However adding an explicit check to spot if kroundup32 has overflowed
is enough to silence it.

NB: this gcc 11.1 issue wasn't visible from within Htslib, but
Samtools tickled it in samtools/dict.c.

Fixes #1283