Remove const from varint_putNN() endp parameters
This works around a spurious warning from GCC 11.1 whereby the compiler
assumed the const pointer cp_end/endp would be read from (because it was
const, so what else would you do with it?). Remove PR #1280's previous
workaround. See <https://gcc.gnu.org/bugzilla/show_bug.cgi?id=100417>.

As endp points to the same array as cp, it might as well be char* as well,
which avoids the spurious warning. As partially noted in htscodecs/varint.h,
the ideal signatures for the getters and putters would be

    int64_t (*varint_get32)(const char **cp, const char *endp, int *err);
    int (*varint_put32)(char *cp, char *endp, int32_t val_p);
    /* etc */

Adding const to the getters would have to occur after htscodecs had
been updated correspondingly (if at all), but happily removing const
from the putter callers can be done in advance of any corresponding
htscodecs change.

While in the neighbourhood, make the "for now" wrappers static.